\name{Example}
\alias{Example}
\title{Example dataset for iECAT}
\description{
Example dataset for iECAT.
}
\format{
  Example contains the following objects:
  \describe{
    \item{Z.list}{a list of 10 genes}
    \item{tbl.external.all.list}{a list of 10 allele count tables of external control samples for the 10 genes in Z.list}
    \item{Y}{a numeric vector of binary phenotypes}
  }
}

