 \name{Open_SSD_wEC}
 \alias{Open_SSD_wEC}
 \title{Open SNP set data (SSD) and external control (EC) info files }
 \description{
 Open a SNP Set data (SSD) and external control (EC) info files. After finishing using the files, please close them by callinsg Close_SSD_wEC function. 
 }
 \usage{
	Open_SSD_wEC(File.SSD, File.Info, File.EC.Info)
 }
\arguments{
      \item{File.SSD}{ name of the SSD file . }
      \item{File.Info}{ name of the SSD info file. }
      \item{File.EC.Info}{ name of the external control (EC) info file. }
}
\value{
 	 a list object of EC.INFO.
}


\author{Seunggeun Lee}

